+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to the project [https://github.com/mazzy-ax/xpoTools](https://github.com/mazzy-ax/xpoTools) will be documented in this file. See also [https://github.com/mazzy-ax/xpoTools/releases](https://github.com/mazzy-ax/xpoTools/releases).
 
+## [1.1.0](https://github.com/mazzy-ax/xpoTools/compare/1.0.1...1.1.0) - 2018-05-03
+
+* To increase a general performance:
+  * moved: the generation of xpp-files is moved from `xpo-import` to `split-xpo` cmdlet
+  * removed: the `xpptext` property from class [xpoItem] is unused and have removed
+
+* Small updated for README and tests of the project meta.
+
 ## [1.0.1](https://github.com/mazzy-ax/xpoTools/compare/1.0.0...1.0.1) - 2018-04-23
 
 * Added object types for ax2012 in `Get-XpoType.ps1`. #1 closed.